+++ src/main/java/org/tinycloud/paginate/dialect/support/GBase8sDialect.java
@@ -13,10 +13,10 @@ public class GBase8sDialect extends AbstractDialect {
      */
     @Override
     public String getPageSql(String originalSql, Page page) {
-        Integer pageNo = page.getPageNum();
-        Integer pageSize = page.getPageSize();
-        int offset = (pageNo - 1) * pageSize;
-        int limit = pageSize;
+        long pageNo = page.getPageNum();
+        long pageSize = page.getPageSize();
+        long offset = (pageNo - 1L) * pageSize;
+        long limit = pageSize;
 
         StringBuilder sql = (new StringBuilder(originalSql)).insert(6, " SKIP " + offset + " FIRST " + limit);
         return sql.toString();
